+++ src/Form/ArticleFormType.php
@@ -16,6 +16,8 @@ use Symfony\Component\Form\FormEvent;
 use Symfony\Component\Form\FormEvents;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
+use Symfony\Component\Validator\Constraints\Image;
+use Symfony\Component\Validator\Constraints\NotNull;
 
 class ArticleFormType extends AbstractType
 {
@@ -31,6 +33,16 @@ class ArticleFormType extends AbstractType
         /** @var Article|null $article */
         $article = $options['data'] ?? null;
         $isEdit = $article && $article->getId();
+        $imageConstraints = [ //we added our constraints to our imageupload file field
+            new Image([
+                'maxSize' => '5M'
+            ])
+        ];
+        if (!$isEdit || !$article->getImageFilename()) { // but in edit (not new) we required an image
+            $imageConstraints[] = new NotNull([
+                'message' => 'Please upload an image',
+            ]);
+        }
 
         $builder
             ->add('title', TextType::class, [
@@ -53,7 +65,8 @@ class ArticleFormType extends AbstractType
             ])
             ->add('imageFile', FileType::class, [
                 'mapped' => false,
-                'required' => false
+                'required' => false,
+                'constraints' => $imageConstraints
             ])
         ;