+++ src/Service/UploaderHelper.php
@@ -1,0 +1,53 @@
+<?php
+
+namespace App\Service;
+
+use Gedmo\Sluggable\Util\Urlizer;
+use Symfony\Component\Asset\Context\RequestStackContext;
+use Symfony\Component\HttpFoundation\File\UploadedFile;
+
+class UploaderHelper
+{
+    const ARTICLE_IMAGE = 'article_image';
+
+    private $uploadsPath;
+    private $requestStackContext;
+
+    /**
+     * Dependency Injection -we need an direction to our uploads path
+     * UploaderHelper constructor.
+     * @param string $uploadsPath
+     */
+    public function __construct(string $uploadsPath, RequestStackContext $requestStackContext)
+    {
+        $this->uploadsPath = $uploadsPath;
+        $this->requestStackContext = $requestStackContext;
+    }
+
+    public function getPublicPath(string $path): string
+    {
+        // needed if you deploy under a subdirectory
+        return $this->requestStackContext
+                ->getBasePath().'/uploads/'.$path;
+    }
+
+    /**
+     * Method to handle uploadingArticleImage
+     * @param UploadedFile $uploadedFile
+     * @return string the name of the file
+     */
+    public function uploadArticleImage(UploadedFile $uploadedFile): string
+    {
+        $destination = $this->uploadsPath.'/'.self::ARTICLE_IMAGE;
+
+        $originalFilename = pathinfo($uploadedFile->getClientOriginalName(), PATHINFO_FILENAME);
+        $newFilename = Urlizer::urlize($originalFilename).'-'.uniqid().'.'.$uploadedFile->guessExtension();
+
+        $uploadedFile->move(
+            $destination,
+            $newFilename
+        );
+
+        return $newFilename;
+    }
+}