+++ src/Twig/AppExtension.php
@@ -3,6 +3,7 @@
 namespace App\Twig;
 
 use App\Service\MarkdownHelper;
+use App\Service\UploaderHelper;
 use Psr\Container\ContainerInterface;
 use Symfony\Contracts\Service\ServiceSubscriberInterface;
 use Twig\Extension\AbstractExtension;
@@ -16,6 +17,17 @@ class AppExtension extends AbstractExtension implements ServiceSubscriberInterfa
     public function __construct(ContainerInterface $container)
     {
         $this->container = $container;
+    }
+
+    /**
+     * We added a new function to get path to our uploaded images
+     * @return TwigFunction[]
+     */
+    public function getFunctions(): array
+    {
+        return [
+            new TwigFunction('uploaded_asset', [$this, 'getUploadedAssetPath'])
+        ];
     }
 
     public function getFilters(): array
@@ -32,10 +44,23 @@ class AppExtension extends AbstractExtension implements ServiceSubscriberInterfa
             ->parse($value);
     }
 
+    /**
+     * We use our service to get path to the uploaded files
+     * @param string $path
+     * @return string
+     */
+    public function getUploadedAssetPath(string $path): string
+    {
+        return $this->container
+            ->get(UploaderHelper::class)
+            ->getPublicPath($path);
+    }
+
     public static function getSubscribedServices()
     {
         return [
             MarkdownHelper::class,
+            UploaderHelper::class, // we aded our service to substribed list of services
         ];
     }
 }