+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Symfony Mailer: Love Sending Emails Again
 ## [26.02.2022]
+- The asset() Function & assets.context
+- URL to Public Assets
+- Upload Field Styling & Bootstrap
+- File Validation
+- Centralizing Upload Logic
 - File Upload Field in a Form
 - Unique (but not Insane) Filenames
 - Where & How to Store the File